FT0: refactoring - merge tasks
TriggerQcTask merged into DigitQcTask (and *Laser)
OutOfBunchCollTask merged into BasicPPTask.cxx
