+++ Modules/FT0/CMakeLists.txt
@@ -7,9 +7,7 @@ target_sources(O2QcFT0 PRIVATE src/TH1ReductorLaser.cxx
                                src/DigitQcTask.cxx
                                src/CFDEffCheck.cxx
                                src/BasicPPTask.cxx
-                               src/OutOfBunchCollTask.cxx
                                src/OutOfBunchCollCheck.cxx
-                               src/TriggerQcTask.cxx
                                src/BasicDigitQcTask.cxx
                                src/ChannelsCheck.cxx
                                src/DigitsCheck.cxx
@@ -38,9 +36,7 @@ add_root_dictionary(O2QcFT0
           include/FT0/DigitQcTaskLaser.h
           include/FT0/DigitQcTask.h
           include/FT0/BasicPPTask.h
-          include/FT0/OutOfBunchCollTask.h
           include/FT0/CFDEffCheck.h
-          include/FT0/TriggerQcTask.h
           include/FT0/OutOfBunchCollCheck.h
           include/FT0/BasicDigitQcTask.h
           include/FT0/ChannelsCheck.h
