+++ Modules/FT0/include/FT0/LinkDef.h
@@ -14,9 +14,7 @@
 //#pragma link C++ class o2::quality_control_modules::ft0::CalibrationTask + ;
 //#pragma link C++ class o2::quality_control_modules::ft0::ChannelTimeCalibrationCheck + ;
 #pragma link C++ class o2::quality_control_modules::ft0::BasicPPTask + ;
-#pragma link C++ class o2::quality_control_modules::ft0::OutOfBunchCollTask + ;
 #pragma link C++ class o2::quality_control_modules::ft0::CFDEffCheck + ;
-#pragma link C++ class o2::quality_control_modules::ft0::TriggerQcTask + ;
 #pragma link C++ class o2::quality_control_modules::ft0::OutOfBunchCollCheck + ;
 #pragma link C++ class o2::quality_control_modules::ft0::RecPointsQcTask + ;
