+++ Modules/FT0/src/DigitQcTaskLaser.cxx
@@ -82,6 +82,44 @@ void DigitQcTaskLaser::rebinFromConfig()
   }
 }
 
+unsigned int DigitQcTaskLaser::getModeParameter(std::string paramName, unsigned int defaultVal, std::map<unsigned int, std::string> choices)
+{
+  if (auto param = mCustomParameters.find(paramName); param != mCustomParameters.end()) {
+    // if parameter was provided check which option was chosen
+    for (const auto& choice : choices) {
+      if (param->second == choice.second) {
+        ILOG(Debug) << "Setting \"" << paramName << "\" to: \"" << choice.second << "\"" << ENDM;
+        return choice.first;
+      }
+    }
+    // param value not allowed - use default but with warning
+    std::string allowedValues;
+    for (const auto& choice : choices) {
+      allowedValues += "\"";
+      allowedValues += choice.second;
+      allowedValues += "\", ";
+    }
+    ILOG(Warning) << "Provided value (\"" << param->second << "\") for parameter \"" << paramName << "\" is not allowed. Allowed values are: " << allowedValues << " setting \"" << paramName << "\" to default value: \"" << choices[defaultVal] << "\"" << ENDM;
+    return defaultVal;
+  } else {
+    // param not provided - use default
+    ILOG(Debug) << "Setting \"" << paramName << "\" to default value: \"" << choices[defaultVal] << "\"" << ENDM;
+    return defaultVal;
+  }
+}
+
+int DigitQcTaskLaser::getNumericalParameter(std::string paramName, int defaultVal)
+{
+  if (auto param = mCustomParameters.find(paramName); param != mCustomParameters.end()) {
+    float val = stoi(param->second);
+    ILOG(Debug) << "Setting \"" << paramName << "\" to: " << val << ENDM;
+    return val;
+  } else {
+    ILOG(Debug) << "Setting \"" << paramName << "\" to default value: " << defaultVal << ENDM;
+    return defaultVal;
+  }
+}
+
 void DigitQcTaskLaser::initialize(o2::framework::InitContext& /*ctx*/)
 {
   ILOG(Info) << "initialize DigitQcTaskLaser" << ENDM; // QcInfoLogger is used. FairMQ logs will go to there as well.
@@ -103,6 +141,35 @@ void DigitQcTaskLaser::initialize(o2::framework::InitContext& /*ctx*/)
   mMapDigitTrgNames.insert({ o2::fit::Triggers::bitLaser, "Laser" });
   mMapDigitTrgNames.insert({ o2::fit::Triggers::bitOutputsAreBlocked, "OutputsAreBlocked" });
   mMapDigitTrgNames.insert({ o2::fit::Triggers::bitDataIsValid, "DataIsValid" });
+
+  mTrgModeThresholdVar = getModeParameter("trgModeThresholdVar",
+                                          TrgModeThresholdVar::kAmpl,
+                                          { { TrgModeThresholdVar::kAmpl, "Ampl" },
+                                            { TrgModeThresholdVar::kNchannels, "Nchannels" } });
+  mTrgModeSide = getModeParameter("trgModeSide",
+                                  TrgModeSide::kAplusC,
+                                  { { TrgModeSide::kAplusC, "A+C" },
+                                    { TrgModeSide::kAandC, "A&C" },
+                                    { TrgModeSide::kA, "A" },
+                                    { TrgModeSide::kC, "C" } });
+  mTrgThresholdTimeLow = getNumericalParameter("trgThresholdTimeLow", -192);
+  mTrgThresholdTimeHigh = getNumericalParameter("trgThresholdTimeHigh", 192);
+  if (mTrgModeSide == TrgModeSide::kAplusC) {
+    mTrgThresholdSCenSum = getNumericalParameter("trgThresholdSCenSum", 300);
+    mTrgThresholdCenSum = getNumericalParameter("trgThresholdCenSum", 600);
+  } else if (mTrgModeSide == TrgModeSide::kAandC) {
+    mTrgThresholdCenA = getNumericalParameter("trgThresholdCenA", 600);
+    mTrgThresholdCenC = getNumericalParameter("trgThresholdCenC", 600);
+    mTrgThresholdSCenA = getNumericalParameter("trgThresholdSCenA", 300);
+    mTrgThresholdSCenC = getNumericalParameter("trgThresholdSCenC", 300);
+  } else if (mTrgModeSide == TrgModeSide::kA) {
+    mTrgThresholdCenA = getNumericalParameter("trgThresholdCenA", 600);
+    mTrgThresholdSCenA = getNumericalParameter("trgThresholdSCenA", 300);
+  } else if (mTrgModeSide == TrgModeSide::kC) {
+    mTrgThresholdCenC = getNumericalParameter("trgThresholdCenC", 600);
+    mTrgThresholdSCenC = getNumericalParameter("trgThresholdSCenC", 300);
+  }
+
   mHistTime2Ch = std::make_unique<TH2F>("TimePerChannel", "Time vs Channel;Channel;Time", sNCHANNELS_PM, 0, sNCHANNELS_PM, 4100, -2050, 2050);
   mHistTime2Ch->SetOption("colz");
   mHistAmp2Ch = std::make_unique<TH2F>("AmpPerChannel", "Amplitude vs Channel;Channel;Amp", sNCHANNELS_PM, 0, sNCHANNELS_PM, 4200, -100, 4100);
@@ -115,10 +182,22 @@ void DigitQcTaskLaser::initialize(o2::framework::InitContext& /*ctx*/)
   }
   mHistOrbitVsTrg = std::make_unique<TH2F>("OrbitVsTriggers", "Orbit vs Triggers;Orbit;Trg", sOrbitsPerTF, 0, sOrbitsPerTF, mMapDigitTrgNames.size(), 0, mMapDigitTrgNames.size());
   mHistOrbitVsTrg->SetOption("colz");
-
+  mHistTriggersSw = std::make_unique<TH1F>("TriggersSoftware", "Triggers from software", mMapDigitTrgNames.size(), 0, mMapDigitTrgNames.size());
+  mHistTriggersSoftwareVsTCM = std::make_unique<TH2F>("TriggersSoftwareVsTCM", "Comparison of triggers from software and TCM;;Trigger name", mMapDigitTrgNames.size(), 0, mMapDigitTrgNames.size(), 4, 0, 4);
+  mHistTriggersSoftwareVsTCM->SetOption("colz");
+  mHistTriggersSoftwareVsTCM->SetStats(0);
   for (const auto& entry : mMapDigitTrgNames) {
     mHistOrbitVsTrg->GetYaxis()->SetBinLabel(entry.first + 1, entry.second.c_str());
+    mHistTriggersSw->GetXaxis()->SetBinLabel(entry.first + 1, entry.second.c_str());
+    mHistTriggersSoftwareVsTCM->GetXaxis()->SetBinLabel(entry.first + 1, entry.second.c_str());
   }
+  mHistTriggersSw->GetXaxis()->SetRange(1, 5);
+  mHistTriggersSoftwareVsTCM->GetXaxis()->SetRange(1, 5);
+  mHistTriggersSoftwareVsTCM->GetYaxis()->SetBinLabel(TrgComparisonResult::kSWonly + 1, "Sw only");
+  mHistTriggersSoftwareVsTCM->GetYaxis()->SetBinLabel(TrgComparisonResult::kTCMonly + 1, "TCM only");
+  mHistTriggersSoftwareVsTCM->GetYaxis()->SetBinLabel(TrgComparisonResult::kNone + 1, "neither TCM nor Sw");
+  mHistTriggersSoftwareVsTCM->GetYaxis()->SetBinLabel(TrgComparisonResult::kBoth + 1, "both TCM and Sw");
+
   mListHistGarbage = new TList();
   mListHistGarbage->SetOwner(kTRUE);
 
@@ -200,6 +279,7 @@ void DigitQcTaskLaser::initialize(o2::framework::InitContext& /*ctx*/)
   // 1-dim hists
   getObjectsManager()->startPublishing(mHistCFDEff.get());
   getObjectsManager()->startPublishing(mHistBC.get());
+  getObjectsManager()->startPublishing(mHistTriggersSw.get());
   // 2-dim hists
   getObjectsManager()->startPublishing(mHistTime2Ch.get());
   getObjectsManager()->setDefaultDrawOptions(mHistTime2Ch.get(), "COLZ");
@@ -215,6 +295,8 @@ void DigitQcTaskLaser::initialize(o2::framework::InitContext& /*ctx*/)
   getObjectsManager()->setDefaultDrawOptions(mHistChDataBits.get(), "COLZ");
   getObjectsManager()->startPublishing(mHistTimeSum2Diff.get());
   getObjectsManager()->setDefaultDrawOptions(mHistTimeSum2Diff.get(), "COLZ");
+  getObjectsManager()->startPublishing(mHistTriggersSoftwareVsTCM.get());
+  getObjectsManager()->setDefaultDrawOptions(mHistTriggersSoftwareVsTCM.get(), "COLZ");
 }
 
 void DigitQcTaskLaser::startOfActivity(Activity& activity)
@@ -231,6 +313,8 @@ void DigitQcTaskLaser::startOfActivity(Activity& activity)
   mHistBCvsFEEmodules->Reset();
   mHistOrbitVsTrg->Reset();
   mHistOrbitVsFEEmodules->Reset();
+  mHistTriggersSw->Reset();
+  mHistTriggersSoftwareVsTCM->Reset();
   for (auto& entry : mMapHistAmp1D) {
     entry.second->Reset();
   }
@@ -272,6 +356,17 @@ void DigitQcTaskLaser::monitorData(o2::framework::ProcessingContext& ctx)
       }
     }
     std::set<uint8_t> setFEEmodules{};
+
+    // reset triggers
+    for (auto& entry : mMapTrgSoftware) {
+      mMapTrgSoftware[entry.first] = false;
+    }
+    float sumAmplA = 0;
+    float sumAmplC = 0;
+    int sumTimeA = 0;
+    int sumTimeC = 0;
+    int nFiredChannelsA = 0;
+    int nFiredChannelsC = 0;
     for (const auto& chData : vecChData) {
       mHistTime2Ch->Fill(static_cast<Double_t>(chData.ChId), static_cast<Double_t>(chData.CFDTime));
       mHistAmp2Ch->Fill(static_cast<Double_t>(chData.ChId), static_cast<Double_t>(chData.QTCAmpl));
@@ -295,6 +390,19 @@ void DigitQcTaskLaser::monitorData(o2::framework::ProcessingContext& ctx)
       }
 
       setFEEmodules.insert(mChID2PMhash[chData.ChId]);
+
+      if (chData.QTCAmpl <= 0)
+        continue;
+
+      if (chData.ChId < mNChannelsA) {
+        sumAmplA += chData.QTCAmpl;
+        sumTimeA += chData.CFDTime;
+        nFiredChannelsA++;
+      } else {
+        sumAmplC += chData.QTCAmpl;
+        sumTimeC += chData.CFDTime;
+        nFiredChannelsC++;
+      }
     }
     if (isTCM) {
       setFEEmodules.insert(mTCMhash);
@@ -303,6 +411,118 @@ void DigitQcTaskLaser::monitorData(o2::framework::ProcessingContext& ctx)
       mHistBCvsFEEmodules->Fill(static_cast<double>(digit.getIntRecord().bc), static_cast<double>(feeHash));
       mHistOrbitVsFEEmodules->Fill(static_cast<double>(digit.getIntRecord().orbit % sOrbitsPerTF), static_cast<double>(feeHash));
     }
+
+    // triggers re-computation
+    mMapTrgSoftware[o2::ft0::Triggers::bitA] = nFiredChannelsA > 0;
+    mMapTrgSoftware[o2::ft0::Triggers::bitC] = nFiredChannelsC > 0;
+
+    int avgTimeA = nFiredChannelsA ? int(sumTimeA / nFiredChannelsA) : 0;
+    int avgTimeC = nFiredChannelsC ? int(sumTimeC / nFiredChannelsC) : 0;
+    int vtxPos = (nFiredChannelsA && nFiredChannelsC) ? (avgTimeC - avgTimeA) / 2 : 0;
+    if (mTrgThresholdTimeLow < vtxPos && vtxPos < mTrgThresholdTimeHigh && nFiredChannelsA && nFiredChannelsC)
+      mMapTrgSoftware[o2::ft0::Triggers::bitVertex] = true;
+
+    // Central/SemiCentral logic
+    switch (mTrgModeSide) {
+      case TrgModeSide::kAplusC:
+        if (mTrgModeThresholdVar == TrgModeThresholdVar::kAmpl) {
+          if (sumAmplA + sumAmplC >= mTrgThresholdCenSum)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (sumAmplA + sumAmplC >= mTrgThresholdSCenSum)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        } else if (mTrgModeThresholdVar == TrgModeThresholdVar::kNchannels) {
+          if (nFiredChannelsA + nFiredChannelsC >= mTrgThresholdCenSum)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (nFiredChannelsA + nFiredChannelsA >= mTrgThresholdSCenSum)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        }
+        break;
+
+      case TrgModeSide::kAandC:
+        if (mTrgModeThresholdVar == TrgModeThresholdVar::kAmpl) {
+          if (sumAmplA >= mTrgThresholdCenA && sumAmplC >= mTrgThresholdCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (sumAmplA >= mTrgThresholdSCenA && sumAmplC >= mTrgThresholdSCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        } else if (mTrgModeThresholdVar == TrgModeThresholdVar::kNchannels) {
+          if (nFiredChannelsA >= mTrgThresholdCenA && nFiredChannelsC >= mTrgThresholdCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (nFiredChannelsA >= mTrgThresholdSCenA && nFiredChannelsC >= mTrgThresholdSCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        }
+        break;
+
+      case TrgModeSide::kA:
+        if (mTrgModeThresholdVar == TrgModeThresholdVar::kAmpl) {
+          if (sumAmplA >= mTrgThresholdCenA)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (sumAmplA >= mTrgThresholdSCenA)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        } else if (mTrgModeThresholdVar == TrgModeThresholdVar::kNchannels) {
+          if (nFiredChannelsA >= mTrgThresholdCenA)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (nFiredChannelsA >= mTrgThresholdSCenA)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        }
+        break;
+
+      case TrgModeSide::kC:
+        if (mTrgModeThresholdVar == TrgModeThresholdVar::kAmpl) {
+          if (sumAmplC >= mTrgThresholdCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (sumAmplC >= mTrgThresholdSCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        } else if (mTrgModeThresholdVar == TrgModeThresholdVar::kNchannels) {
+          if (nFiredChannelsC >= mTrgThresholdCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitCen] = true;
+          if (nFiredChannelsC >= mTrgThresholdSCenC)
+            mMapTrgSoftware[o2::ft0::Triggers::bitSCen] = true;
+        }
+        break;
+    }
+
+    for (const auto& entry : mMapTrgSoftware)
+      if (entry.second)
+        mHistTriggersSw->Fill(entry.first);
+
+    for (const auto& entry : mMapTrgSoftware) {
+      bool isTCMFired = digit.mTriggers.getTriggersignals() & (1 << entry.first);
+      bool isSwFired = entry.second;
+      if (!isTCMFired && isSwFired)
+        mHistTriggersSoftwareVsTCM->Fill(entry.first, TrgComparisonResult::kSWonly);
+      else if (isTCMFired && !isSwFired)
+        mHistTriggersSoftwareVsTCM->Fill(entry.first, TrgComparisonResult::kTCMonly);
+      else if (!isTCMFired && !isSwFired)
+        mHistTriggersSoftwareVsTCM->Fill(entry.first, TrgComparisonResult::kNone);
+      else if (isTCMFired && isSwFired)
+        mHistTriggersSoftwareVsTCM->Fill(entry.first, TrgComparisonResult::kBoth);
+
+      if (isTCMFired != isSwFired) {
+        // (*) = triggers.amplA/C are sums of amplitudes **divided by 8**
+        auto msg = Form(
+          "Software does not reproduce TCM decision! \n \
+                         trigger name: %s\n \
+                         TCM / SW: \n \
+                         hasFired   = %d / %d \n \
+                         nChannelsA = %d / %d \n \
+                         nChannelsC = %d / %d \n \
+                         sumAmplA   = %d / %d (*) \n \
+                         sumAmplC   = %d / %d (*) \n \
+                         timeA      = %d / %d \n \
+                         timeC      = %d / %d \n \
+                         vertexPos  = -- / %d",
+          mMapDigitTrgNames[entry.first].c_str(),
+          isTCMFired, isSwFired,
+          digit.mTriggers.getNChanA(), nFiredChannelsA,
+          digit.mTriggers.getNChanC(), nFiredChannelsC,
+          digit.mTriggers.getAmplA(), int(sumAmplA / 8),
+          digit.mTriggers.getAmplC(), int(sumAmplC / 8),
+          digit.mTriggers.getTimeA(), avgTimeA,
+          digit.mTriggers.getTimeC(), avgTimeC, vtxPos);
+        ILOG(Debug, Support) << msg << ENDM;
+      }
+    }
+    // end of triggers re-computation
   }
 }
 
@@ -333,7 +553,8 @@ void DigitQcTaskLaser::reset()
   mHistBCvsFEEmodules->Reset();
   mHistOrbitVsTrg->Reset();
   mHistOrbitVsFEEmodules->Reset();
-
+  mHistTriggersSw->Reset();
+  mHistTriggersSoftwareVsTCM->Reset();
   for (auto& entry : mMapHistAmp1D) {
     entry.second->Reset();
   }
